reusue implemenation for constructor by using this
