+++ TdLib/Bindings/Receiver.cs
@@ -9,14 +9,6 @@ using Newtonsoft.Json;
 
 namespace TdLib.Bindings
 {
-    public interface IReceiver: IDisposable
-    {
-        event EventHandler<TdApi.Object> Received;
-        event EventHandler<TdApi.AuthorizationState> AuthorizationStateChanged;
-        event EventHandler<Exception> ExceptionThrown;
-
-        void Start();
-    }
     public sealed class Receiver(ITdJsonClient tdJsonClient, TimeSpan receiverTimeOut) : IReceiver
     {
         private readonly Converter _converter = new();
